lf inline: lfops: move prune command under wt subcommand
Reorganizes `lfops prune` to `lfops wt prune` for better command grouping. Deletes the standalone prune module and integrates the functionality into wt.py.
