+++ src/loopflow/lfops/commands.py
@@ -11,7 +11,6 @@ from loopflow.lfops import commit as commit_module
 from loopflow.lfops import rebase as rebase_module
 from loopflow.lfops import summarize as summarize_module
 from loopflow.lfops import sync as sync_module
-from loopflow.lfops import prune as prune_module
 from loopflow.lfops import wt as wt_module
 
 app = typer.Typer(help="Loopflow operations")
@@ -24,7 +23,6 @@ commit_module.register_commands(app)
 rebase_module.register_commands(app)
 summarize_module.register_commands(app)
 sync_module.register_commands(app)
-prune_module.register_commands(app)
 wt_module.register_commands(app)
